Moved other classes to their own namespace
