+++ src/ContaoCommunityAlliance/Composer/Plugin/Installer/CopyInstaller.php
@@ -13,7 +13,7 @@
  * @license LGPL-3.0+
  */
 
-namespace ContaoCommunityAlliance\Composer\Plugin;
+namespace ContaoCommunityAlliance\Composer\Plugin\Installer;
 
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
