+++ src/ContaoCommunityAlliance/Composer/Plugin/Installer/SymlinkInstaller.php
@@ -13,7 +13,7 @@
  * @license LGPL-3.0+
  */
 
-namespace ContaoCommunityAlliance\Composer\Plugin;
+namespace ContaoCommunityAlliance\Composer\Plugin\Installer;
 
 use Composer\Package\PackageInterface;
