+++ src/ContaoCommunityAlliance/Composer/Plugin/Plugin.php
@@ -31,10 +31,14 @@ use Composer\Script\ScriptEvents;
 use Composer\Script\PackageEvent;
 use Composer\Package\LinkConstraint\EmptyConstraint;
 use Composer\Package\LinkConstraint\VersionConstraint;
+use ContaoCommunityAlliance\Composer\Plugin\Dependency\ConfigManipulator;
 use ContaoCommunityAlliance\Composer\Plugin\Environment\ContaoEnvironmentFactory;
 use ContaoCommunityAlliance\Composer\Plugin\Environment\ContaoEnvironmentInterface;
 use ContaoCommunityAlliance\Composer\Plugin\Exception\ConstantsNotFoundException;
 use ContaoCommunityAlliance\Composer\Plugin\Exception\DuplicateContaoException;
+use ContaoCommunityAlliance\Composer\Plugin\Installer\CopyInstaller;
+use ContaoCommunityAlliance\Composer\Plugin\Installer\RunonceManager;
+use ContaoCommunityAlliance\Composer\Plugin\Installer\SymlinkInstaller;
 use RuntimeException;
 
 /**
